Additional reformatting and refactoring
