+++ src/clients/common/progressbar.js
@@ -1,0 +1,14 @@
+//-----------------------------------------------------------------------------
+// Functions for progress indicator bar
+//-----------------------------------------------------------------------------
+
+export function startProcess() {
+    const progressbar = document.querySelector(`#progressbar`);
+    progressbar.className = 'progress-bar';
+  }
+  
+  export function stopProcess() {
+    const progressbar = document.querySelector(`#progressbar`);
+    progressbar.className = 'progress-bar-inactive';
+  }
+
