+++ src/clients/common/stylesheets/common.css
@@ -1,6 +1,7 @@
-@import "/common/stylesheets/generics.css";
-@import "/common/stylesheets/navbar.css";
-@import "/common/stylesheets/snackbar.css";
-@import "/common/stylesheets/marc.css";
 @import "/common/stylesheets/components.css";
+@import "/common/stylesheets/generics.css";
 @import "/common/stylesheets/login.css";
+@import "/common/stylesheets/marc.css";
+@import "/common/stylesheets/navbar.css";
+@import "/common/stylesheets/progressbar.css";
+@import "/common/stylesheets/snackbar.css";
