+++ src/clients/common/stylesheets/components.css
@@ -578,67 +578,3 @@ dialog:modal {
   padding-left: 9px;
   padding-right: 9px;
 }
-
-/* ---------------------------------------------------------------------------- */
-/* Loading bar */
-/* ---------------------------------------------------------------------------- */
-
-.progress-bar-inactive {
-  height: 3px;
-  width: 100%;
-  margin: 0;
-}
-
-.progress-bar,
-.progress-bar:before {
-  height: 3px;
-  width: 100%;
-  margin: 0;
-}
-
-.progress-bar {
-  background-color: #b3d4fc;
-  display: -webkit-flex;
-  display: flex;
-}
-
-.progress-bar:before {
-  background-color: #3f51b5;
-  content: '';
-  animation: running-progress 2s linear infinite;
-  -webkit-animation: running-progress 2s linear infinite;
-}
-
-@keyframes running-progress {
-  0% {
-    margin-left: 0px;
-    margin-right: 100%;
-  }
-
-  50% {
-    margin-left: 0%;
-    margin-right: 0%;
-  }
-
-  100% {
-    margin-left: 100%;
-    margin-right: 0%;
-  }
-}
-
-@-webkit-keyframes running-progress {
-  0% {
-    margin-left: 0px;
-    margin-right: 100%;
-  }
-
-  50% {
-    margin-left: 0%;
-    margin-right: 0%;
-  }
-
-  100% {
-    margin-left: 100%;
-    margin-right: 0%;
-  }
-}
